fix: relative path for log

+++ lib/routes2spec/command.rb
@@ -39,26 +39,27 @@ module Routes2spec
 
       results = inspector.format(formatter, routes_filter)
       results.each do |result|
-        outfile = Rails.root.join("spec/requests", *result[:namespaces], "#{result[:name].underscore}_spec.rb")
+        relative_path = File.join("spec/requests", *result[:namespaces], "#{result[:name].underscore}_spec.rb")
+        outfile = Rails.root.join(relative_path)
         FileUtils.mkdir_p(File.dirname(outfile))
 
         unless File.exist?(outfile)
-          Routes2spec.log "Generating: #{outfile}"
+          Routes2spec.log "Generating: #{relative_path}"
           File.write(outfile, result[:content], mode: "w")
           next
         end
 
         if options.force_overwrite?
-          Routes2spec.log "Overwriting: #{outfile}"
+          Routes2spec.log "Overwriting: #{relative_path}"
           File.write(outfile, result[:content], mode: "w")
           next
         end
 
-        Routes2spec.log "Already exists: #{outfile}"
-        if options.overwrite?
-          if file_collision(outfile)
-            File.write(outfile, result[:content], mode: "w")
-          end
+        Routes2spec.log "Already exists: #{relative_path}"
+
+        if options.overwrite? && file_collision(relative_path)
+          say "Overwriting..."
+          File.write(outfile, result[:content], mode: "w")
         end
       end
     end
